Fixed a MySQL error that occurred when deleting employees that were never scheduled.  Head only as this wasn't an issue in 14.3. Reviewers:  Allen, Jason
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@11805 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/Employees.cs
@@ -141,8 +141,10 @@ namespace OpenDentBusiness{
 				DeletedObjects.SetDeleted(DeletedObjectType.ScheduleProv,PIn.Long(table.Rows[i]["ScheduleNum"].ToString()));
 				listScheduleNums.Add(table.Rows[i]["ScheduleNum"].ToString());
 			}
-			command="DELETE FROM scheduleop WHERE ScheduleNum IN("+POut.String(String.Join(",",listScheduleNums))+")";
-			Db.NonQ(command);
+			if(listScheduleNums.Count>0) {
+				command="DELETE FROM scheduleop WHERE ScheduleNum IN("+POut.String(String.Join(",",listScheduleNums))+")";
+				Db.NonQ(command);
+			}
 			//command="DELETE FROM scheduleop WHERE ScheduleNum IN(SELECT ScheduleNum FROM schedule WHERE EmployeeNum="+POut.Long(employeeNum)+")";
 			//Db.NonQ(command);
 			command="DELETE FROM schedule WHERE EmployeeNum="+POut.Long(employeeNum);
